Extend SMTP interface and map from generic custom fields interface and map
